change bankId to short in Account entity class

+++ src/main/java/com/idehalborz/expensetracker/entity/Account.java
@@ -18,6 +18,6 @@ public class Account {
     private int id;
     private String name;
     private String accountNumber;
-    private byte bankId;
+    private short bankId;
     private byte accountType;
 }
